Updater now uses appropriate box art in the wizard watermarks depending on which Game it is.
