+++ res/fa/updater/forged_alliance_watermark.png
