+++ src/fa/__init__.py
@@ -20,8 +20,6 @@
 # Initialize logging system
 import logging
 
-from fa.path import loadPathSC, loadPath
-
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
@@ -32,18 +30,12 @@ DEFAULT_LIVE_REPLAY = True
 DEFAULT_RECORD_REPLAY = True
 DEFAULT_WRITE_GAME_LOG = False
 
-
 # We only want one instance of Forged Alliance to run, so we use a singleton here (other modules may wish to connect to its signals so it needs persistence)
 from process import instance as instance
 from play import play as play
 
+from fa.path import gamepath
 
-# This is the game path, a string pointing to the player's actual install of Forged Alliance
-gamepath = None
-
-# Initial Housekeeping
-loadPath()
-loadPathSC()
 
 import check
 import maps
@@ -53,3 +45,4 @@ import proxies
 import updater
 import upnp
 import faction
+
